fixing type inference
